Move the 128256-wide topk cell to L2 nightly instead of skipping it on ttsim
Replaces the skip_wide_topk_on_sim marker with a relocation, matching what #55699
did for the reduce group's other ttsim-pathological shapes.

test_large_2d_topk now lives in
tests/ttnn/nightly/unit_tests/operations/reduction/test_reduce_large_shapes.py and
delegates to test_2d_topk, per that file's convention of sharing the sanity
assertions so they cannot drift. Its bfloat16 branch selects exactly the thresholds
the moved test hardcoded, so behaviour is unchanged.

The nightly reduction group runs on wh_n300_civ2 + bh_p100 (and wh_n150_civ2 +
bh_p150b_civ2 for the release group), none of which are simulator SKUs, so the cell
leaves ttsim without needing a marker. The small test_topk_fp32_uint32_indices stays
in sanity and keeps the UInt32 index branch covered there.
